ChatBox logic seperated to hook and used seperated content and chatbox now
