+++ src/components/chat-box/ChatBox.tsx
@@ -1,34 +1,23 @@
-import React, { useEffect, useState } from "react";
-
-import { useGeminiResponse } from "../../hooks/useGeminiResponse";
-import {
-  removeLocalStorageData,
-  loadChatData,
-  saveChatData,
-} from "../../utils/dataUtils";
-import { Message, Sender } from "../../types/messageType";
+import React from "react";
+import { useChatWithAi } from "../../hooks/useChatWithAi";
 import "./style.scss";
-import { MessageLine } from "../message-line/MessageLine";
-import {
-  abortCurrentPrompt,
-  resetSession,
-} from "../../utils/fetchGeminiResponse";
 import { TextGenerateEffectFx } from "../ui/fx/textGenerateEffectFx";
 import { VanishInputFx } from "../ui/fx/vanishInputFx";
+import { MessageLine } from "../message-line/MessageLine";
 import { BackgroundBeamsFx } from "../ui/fx/backgroundBeamsFx";
+import { Sender } from "@/types/messageType";
 
 const ChatBox: React.FC = () => {
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [input, setInput] = useState<string>("");
-  const [latestAIMessageIndex, setLatestAIMessageIndex] = useState<
-    number | null
-  >(null);
-
   const {
+    messages,
     loading,
-    messages: fetchedMessages,
-    fetchResponse,
-  } = useGeminiResponse();
+    latestAIMessageIndex,
+    handleInputChange,
+    handleSubmit,
+    clearChatHistory,
+    abortCurrentPrompt,
+    resetSession,
+  } = useChatWithAi("chatbox");
 
   const placeholders = [
     "Type your interest...",
@@ -37,45 +26,12 @@ const ChatBox: React.FC = () => {
     "Let's create a content!",
   ];
 
-  useEffect(() => {
-    loadChatData(setMessages);
-  }, []);
-
-  useEffect(() => {
-    if (fetchedMessages.length) {
-      setMessages((prevMessages) => [...prevMessages, ...fetchedMessages]);
-      setLatestAIMessageIndex(messages.length);
-    }
-  }, [fetchedMessages]);
-
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setInput(event.target.value);
-  };
-
-  const sendMessage = () => {
-    if (!input.trim()) {
-      return;
-    }
-
-    const userMessage = { sender: Sender.USER, text: input };
-    setMessages((prevMessages) => [...prevMessages, userMessage]);
-    saveChatData(userMessage);
-    fetchResponse(input, "chatbox");
-    setInput("");
-  };
-
-  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    sendMessage();
-  };
-
   return (
     <div className="chatbox-container">
       <div className="chat-messages">
         {messages.map((message, index) => (
           <div key={index} className={`message ${message.sender}`}>
             {message.sender === Sender.AI ? (
-              // Apply TextGenerateEffect only to the latest AI message
               index === latestAIMessageIndex ? (
                 <TextGenerateEffectFx
                   words={message.text || ""}
@@ -83,12 +39,9 @@ const ChatBox: React.FC = () => {
                   filter={false}
                 />
               ) : (
-                // TODO !!! correct style match with ai response with effect
-                // Render older AI messages with static MessageLine
                 <MessageLine text={message.text} />
               )
             ) : (
-              // Render user messages with MessageLine
               <MessageLine text={message.text} />
             )}
           </div>
@@ -104,45 +57,10 @@ const ChatBox: React.FC = () => {
         />
       </div>
 
-      <div
-        style={{
-          display: "flex",
-          flexDirection: "row",
-          width: "100%",
-          justifyContent: "space-around",
-          alignContent: "space-around",
-          fontSize: "8px",
-          zIndex: "20",
-          height: "40px",
-        }}
-      >
-        <button
-          style={{
-            color: "white",
-          }}
-          onClick={() => {
-            removeLocalStorageData("chatHistory", () => setMessages([]));
-            setLatestAIMessageIndex(null);
-          }}
-        >
-          Clear Chat History
-        </button>
-        <button
-          style={{
-            color: "white",
-          }}
-          onClick={abortCurrentPrompt}
-        >
-          Stop Running Prompt
-        </button>
-        <button
-          style={{
-            color: "white",
-          }}
-          onClick={resetSession}
-        >
-          Reset AI Session
-        </button>
+      <div className="chat-actions">
+        <button onClick={clearChatHistory}>Clear Chat History</button>
+        <button onClick={abortCurrentPrompt}>Stop Running Prompt</button>
+        <button onClick={resetSession}>Reset AI Session</button>
       </div>
 
       <BackgroundBeamsFx />
