+++ src/components/ui/fx/backgroundBeamsFx.tsx
@@ -60,7 +60,7 @@ export const BackgroundBeamsFx = React.memo(
     return (
       <div
         className={cn(
-          "absolute  h-full w-full inset-0  [mask-size:40px] [mask-repeat:no-repeat] flex items-center justify-center",
+          "absolute  h-full w-full inset-0  [mask-size:40px] [mask-repeat:no-repeat] flex items-center justify-center pointer-events-none",
           className
         )}
       >
